Roll back home-assistant to 2025.6.3
HASS broke MQTT in 2025.7 with home-assistant/core#146722. Apparently
warnings were supposed to have been logged for 6 months, but I'm not
finding anything in my logs and don't recall seeing a warning in the
web UI.

Supposedly this is fixed by home-assistant/core#149006 which looks
like it'll be in 2025.7.3. Until then, pin HASS to 2025.6.3.