Default ndots to 1 for non-kube-system pods
Upgraded the cluster to 1.33 and enabled the MutatingAdmissionPolicy
feature gates and APIs. This allows me to default ndots to 1 in pods
so I don't have to troubleshoot weird dns failures. It's not perfect
because JSONPatch is super limited in what it can do, but we're locked
into using it because ApplyPatch won't touch atomic lists and sadly
.spec.dnsConfig.options is marked as atomic.